chore: Use moby for GPU instances
This was stuck on docker 1.13 due to packaging issues with nvidia-docker
(which has a hard dep on docker specific package versions).

This change manually installs nvidia-docker2, but uses apt to install
the rest of the nvidia-* packages which don't have this hard dep.

nvidia-docker2 is really just a docker CLI clone (not sure if there are
changes to it) and a custom `/etc/docker/daemon.json` which sets the
default runtime to nvidia-container-runtime.