fix: Windows + VMSS external svc routing broken (#1028)
* fix: "Windows + VMSS external svc routing broken"

The VMSSs for Linux and Windows agentpools follow a different
naming patterns. However, when populating azure.json, aks-engine
does not take that into account and we end up with the wrong name
for the primareyScaleSetName variable. Thus, the LB created for
service will have a empty backend pool, as there isn't a
VMSS with the desired name.

Properly populating the ARM variable for the primaryScaleSet will
fix this issue.

Fixes: #809

* chore: fix unit tests, protect against nil dereference

* chore: more tests, get rid of dead code, rationalize GetPrimaryAvailabilitySetName

* chore: lint

* test: fix unit tests