fix: Ensure etcd starts after disk is mounted (#1515)
There is currenlty no ordering specified that the disk
`/var/lib/etcddisk` is available before etcd starts.

Predominately this is the case but on a hypervisor crash that disk
gets fscked and may not be available until later. Etcd will try for a
little while but will give up eventually leaving the master node
offline :-(

This patch tells systemd to add the ordering commands such that
etcd.service won't start till the disk is mounted.