feat: upgrade uses keyvault to create kubeconfig
This commit adds a `GetKeyVaultSecret` function to the `AKSEngineClient`
interface. This is used in [cmd/upgrade.go](./cmd/upgrade.go) to fetch
the caCrt, kubectlCrt, and kubectlKey from Azure KeyVault and populate a
kubeconfig, which it then uses to roll a kubernetes cluster.

As this is a WIP PR, it is probably unnecessary to point out that I've
made some not-ideal design decisions in getting this to solve my use
case. The first that comes to mind is how I opted to log a warning
whenever a KeyVault request is attempted but fails. I wasn't sure what
the desired behavior was there. Perhaps it should just fail out.

I can't pretend I am a golang developer. What I present here is a hack
job, and I'm sure I broke several golang conventions along the way. One
part of the code that comes to mind is my instantiation of an
`api.CertificateProfile` struct to use as a temporary placeholder while
I manipulate the properties of the uc `UpgradeCommand` object. I do this
because the `GenerateKubeConfig` function takes the entire `Properties`
structure, and I only required 3 properties from it.

I will follow up with a second, much smaller PR, that shows how I worked
around the same limitation by simply reading a kubeconfig file from a
path that I pass in from the command line.