feat: Adds kubeconfig option to upgrade cmd
This PR accounts for my second strategy to solve the limitation
described in #2396. Rather than introduce the KeyVault client, so we can
run `GetSecret` and fetch our certs and keys, I read a local config file
based on what's passed into the command line.

This is not a complete PR, because I think the same stuff would need to
be done anywhere else `GenerateKubeConfig` is called.