Set the csi-secrets-store to have a priority class
Daemonsets are needed to run on all nodes but in a busy cluster with dynamic scaling, it is possible to have enough pending pods from replica sets that a new node fills up before the daemonset pod is ready to be scheduled.  Without priority classes, this can cause daemonsets to be pending to the pod they are bound to and can not run elsewhere.

In this change I added a priority class and use it in the daemonsets.  We could look at using the priority class system-node-critical which is reserved for kubernetes system daemonsets.  I don't think it is valid to use that here.

Just having a priority higher than 0 will usually address this as 99.99% of all workload pods are 0 (or below zero for the easily evicted buffer pods and such)