Add fix/workaround for DNS TCP issues in the clusters
We update the /opt/azure/containers/kubelet.sh which is the pre-start
hook for the kubelet service in systemd.  It already does some iptables
work due to what it needed so we add in the following two lines (in
order):

iptables -t security -D OUTPUT -d 168.63.129.16/32 -p tcp -m tcp --dport 53 -j ACCEPT || true
iptables -t security -I OUTPUT 1 -d 168.63.129.16/32 -p tcp -m tcp --dport 53 -j ACCEPT

The two lines are (1) to delete the rule if it is there already and
then (2) to add the rule at the top (first entry) which specifically
allows TCP to the magic VM host controlled IP address for DNS (which is
the upstream DNS provider)

Why the delete?  Well, if the kubelet gets restarted, we don't want to
add the rule again.  The -I will add the rule again, even if it is there
already.  So, we delete the rule first.  It also makes sure that the
rule is added at the top of the chain such that this will also
effectively move it to the top if it was already there.  In other words,
it just makes sure that the rule exists and is at the right spot.  If it
does not exist, the delete (-D) will "fail" to delete the non-existing
rule, which is the end state we want so we don't care if the delete
failed and thus the "|| true".  Then the second step is to insert the
rule at the top of the chain, thus making sure that no other rule blocks
access before it can do the ACCEPT

This fix works around a bug in the Azure Linux Agent (walinux) which
block all access to 168.63.129.16 from non-root processes in order to
protect the MSI token distribution mechanism from non-privileged
processes. However, the blanket rule also blocks the upstream TCP DNS
requests. Normally DNS is UDP but there are cases where TCP must be
used, especially when the DNS record returned is larger than the
datagram size limit on the UDP request. Blocking this thus has very
negative impacts for certain DNS queries due to the record size
(especially with the complex structures that get formed by things like
traffic manager)

There is an old, long standing issue opened on the Azure linux agent
about this: https://github.com/Azure/WALinuxAgent/issues/1673