Initial support work for Ubuntu 20.04
While we are in the process of moving to Cluster API, we have a large
number of services running in many clusters and it is unlikely that we
will have completed the Cluster API integration, validation, and
migration before the Ubuntu 18.04 end of support date of April 2023.
As such, this is a stop-gap to allow us to migrate up to Ubuntu 20.04
while we are continuing our work towards Cluster API.

This is a very preliminary checkin.  The clusters now do come up if you
use the ubuntu-20.04 distro name - but it uses a generic ubuntu-20.04
image and thus is not at all zero egress yet.

This includes a number of minor changes needed to get them working,
including a change to deal with a difference as to how cloud-init runs
in 20.04 with IP address assignment.

Again, the goal is to be as similar to 18.04 as we can be to minimize
the risks for a short term gap update until we can complete Cluster API
migration.

Oh, and some tests don't work yet but that will be fixed after this
initial commit.